Fixes for redirects and subclassing.
The `_edit()` and `save()` methods weren't returning their promise
object. This causes problems if a subclass extends these methods and
attempts to use `super(...)then(...)`.

When the edit and delete methods get invoked with an Ajax request from
Polymer elements, they need to return a response, rather than do a
redirect. It seems reasonable to subclass `EditController` and then
extend the methods to do whatever handshake is appropriate. This won't
work if the method has already done a `redirect()`. Changed the code to
skip the redirect if `redirectAfterEdit`, `redirectAfterCreate` or
`redirectAfterDelete` are empty, so the subclass can suppress the
redirect.