Revised approch for disabling redirect.
In `EditController.js`, added a `redirect` option (true by default) that
can be set `false` to disable redirect. (Instead of checking for empty
`redirectAfterEdit`, `redirectAfterCreate` and `redirectAfterDelete`,
which was going to break existing code.)

This doesn't allow redirects to be disabled selectively, but that seemed
like overkill. (The idea being you get to use the traditional redirect
approach, or something different like Polymer modal forms, but not a
mix.)